Authorize /lgtm and /approve with OWNERS file
When an OWNERS file is present, the /lgtm and /approve commands are
authorized using the contents of the file instead of org/repo
membership. The github action does not checkout the repository itself to
locate OWNERS. The workflow is responsible for checking out the repo
first before calling the prow action.

The format of the OWNERS file is a subset of prow's OWNERS file.
I am only looking for the file at the root of the repository.
The format is as follows:

```yaml
approvers:
- user1
- user2

reviewers:
- user1
- user2
- user3
```

/lgtm checks if the commentor is in the reviewers section, and /approve
checks if the commentor is in the approvers section. Membership in
approvers does not authorize the use of /lgtm.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>