Fix main test when run for a PR
The main test relies on the current GitHub event name, and that is
populated via the environment variable GITHUB_EVENT_NAME.

When you run locally, that's unset, and the test passes.

When you run the workflow after pushing to main, it's set to "push",
which isn't handled by main.ts, so it hits the default case and returns
"not yet supported".

But when the workflow is run in a PR, that env var is set to
"pull_request" which is an event that main tries to handle. So the test
fails in an unexpected way (throwing an error) instead of printing "not
yet supported".

I have updated the test to clear out GITHUB_EVENT_NAME so that we get
the same test behavior in all cases.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>